renamed get_key.py and turned it into a class
